I have moved images from ./public to ./src/Assets. In ./public/ is only two images left that concern to README file.
