+++ src/Assets/Images/logoVector.png
